+++ app/src/main/java/com/example/captstone/viewPagerAdapters/HomeViewPagerAdapter.java
@@ -1,15 +1,14 @@
-package com.example.captstone;
+package com.example.captstone.viewPagerAdapters;
 
 import androidx.annotation.NonNull;
 import androidx.fragment.app.Fragment;
 import androidx.viewpager2.adapter.FragmentStateAdapter;
 
 import com.example.captstone.fragments.ReviewFeedFragment;
-import com.example.captstone.fragments.TrendingFeedFragment;
 
-public class TrendingViewPagerAdapter extends FragmentStateAdapter {
+public class HomeViewPagerAdapter extends FragmentStateAdapter {
 
-    public TrendingViewPagerAdapter(@NonNull Fragment fragment) {
+    public HomeViewPagerAdapter(@NonNull Fragment fragment) {
         super(fragment);
     }
 
@@ -18,14 +17,16 @@ public class TrendingViewPagerAdapter extends FragmentStateAdapter {
     public Fragment createFragment(int position) {
         switch (position) {
             case 0:
-                return new TrendingFeedFragment();
+                return new ReviewFeedFragment();
+            case 1:
+                return new ReviewFeedFragment();
             default:
-                return new TrendingFeedFragment();
+                return new ReviewFeedFragment();
         }
     }
 
     @Override
     public int getItemCount() {
-        return 1;
+        return 2;
     }
 }
